Implement arrays
Fixes #2
Remove JSONWrapper setValue as it is no longer needed
Remove setting.cpp
Add a ConnectionID to SettingData so they can be compared easier
Add methods for deep merging objects/arrays in rapidjson
Clean up the way the SettingManager and its document is stored
The concept of a SettingManager being "loaded" doesn't exist for now
Add ostream operator<< for our custom Object and Array types
Add the concept of "paths" in Settings. This path is used for JSON
Pointers.
Make use of our signal library by adding a "valueChanged" signal to
SettingData
Setting should probably keep a reference to it that "outside people" can
use.
SettingData no longer has a concept of a parent. This is all handled
with JSON Pointers now
Change permissions in SettingData to avoid confusion in what can be used
outside the library
Add method for pretty printing document to stdout