Add new setting option: Remote
This is specified for any settings that should only be stored as a
setting but never actually saved locally.
This allows for users to implement their own multi-application settings.
One application only registers settings and levererages the settings
system to notify the other aplication of the settings it has loaded, and
gets the data over network.