Add new test file for "advanced" (de)serialization
The SettingManager instance is now stored as a shared_ptr instead of by value
Allow multiple SettingManager instances to be used. There are still a few functions that cannot be run on the non-main instance
Static methods that have a non-static counterpart have been prefixed with `g`

Add some array/container helper functions to Setting and SettingData. Namely getArray and push_back

getArray can be used to iterate over a setting that holds a container
push_back can be used to push an object back into the container that the setting holds
push_back needs to also be able to take a const ref, but that's for later